Improve `stb serve`
This used the older syntax for compiled assets, which is now replaced
with `Project.compiled_assets`. This improves which paths are "watched"
for rebuilds and browser reloads by `sphinx-autobuild`. Additionally, a
clear separation of `sphinx-autobuild` flags and Sphinx flags aids
readability.

Add a diagnostic error for situations where `sphinx-autobuild` crashes
or exits with a not-clean exit code.