Use the `node` from PATH, if it matches the required version
This is to aid re-distributors who wish to avoid fetching NodeJS from
an arbitrary URL. Note that themes specify the required NodeJS version
in their pyproject.toml file, upto the patch version. Any mismatch will
result in a build to execute with NodeJS binary fetched from
https://nodejs.org/download/release/ or being compiled from source if
run on a platform that is not supported.